Use filenames matching class names for actions/transforms
