+++ include/fullscore/actions/transforms/ClearMeasure.h
