+++ include/fullscore/actions/transforms/DoubleDuration.h
