+++ include/fullscore/actions/transforms/EraseNote.h
