+++ include/fullscore/actions/transforms/HalfDuration.h
