+++ include/fullscore/actions/transforms/RemoveDot.h
