+++ src/actions/transforms/AddDot.cpp
@@ -2,7 +2,7 @@
 
 
 
-#include <fullscore/actions/transforms/add_dot_action.h>
+#include <fullscore/actions/transforms/AddDot.h>
 
 #include <fullscore/transforms/add_dot_transform.h>
 #include <fullscore/models/note.h>
