+++ src/actions/transforms/ClearMeasure.cpp
@@ -1,7 +1,7 @@
 
 
 
-#include <fullscore/actions/transforms/clear_measure_action.h>
+#include <fullscore/actions/transforms/ClearMeasure.h>
 
 #include <fullscore/transforms/clear_measure_transform.h>
 #include <fullscore/models/note.h>
