+++ src/actions/transforms/DoubleDuration.cpp
@@ -2,7 +2,7 @@
 
 
 
-#include <fullscore/actions/transforms/double_duration_action.h>
+#include <fullscore/actions/transforms/DoubleDuration.h>
 
 #include <fullscore/transforms/double_duration_transform.h>
 #include <fullscore/models/note.h>
