+++ src/actions/transforms/EraseNote.cpp
@@ -2,7 +2,7 @@
 
 
 
-#include <fullscore/actions/transforms/erase_note_action.h>
+#include <fullscore/actions/transforms/EraseNote.h>
 
 #include <fullscore/models/note.h>
 #include <fullscore/transforms/erase_note_transform.h>
