+++ src/actions/transforms/HalfDuration.cpp
@@ -2,7 +2,7 @@
 
 
 
-#include <fullscore/actions/transforms/half_duration_action.h>
+#include <fullscore/actions/transforms/HalfDuration.h>
 
 #include <fullscore/models/note.h>
 #include <fullscore/transforms/half_duration_transform.h>
