+++ src/actions/transforms/Invert.cpp
@@ -2,7 +2,7 @@
 
 
 
-#include <fullscore/actions/transforms/invert_action.h>
+#include <fullscore/actions/transforms/Invert.h>
 
 #include <fullscore/models/note.h>
 #include <fullscore/transforms/invert_transform.h>
