+++ src/actions/transforms/Retrograde.cpp
@@ -2,7 +2,7 @@
 
 
 
-#include <fullscore/actions/transforms/retrograde_action.h>
+#include <fullscore/actions/transforms/Retrograde.h>
 
 #include <fullscore/transforms/retrograde_transform.h>
 #include <fullscore/models/note.h>
