+++ src/actions/transforms/ToggleRest.cpp
@@ -2,7 +2,7 @@
 
 
 
-#include <fullscore/actions/transforms/toggle_rest_action.h>
+#include <fullscore/actions/transforms/ToggleRest.h>
 
 #include <fullscore/transforms/toggle_rest_transform.h>
 #include <fullscore/models/note.h>
