+++ src/actions/transforms/TransposeUp.cpp
@@ -2,7 +2,7 @@
 
 
 
-#include <fullscore/actions/transforms/transpose_down_action.h>
+#include <fullscore/actions/transforms/TransposeUp.h>
 
 #include <fullscore/models/note.h>
 #include <fullscore/transforms/transpose_transform.h>
@@ -10,28 +10,28 @@
 
 
 
-Action::Transform::TransposeDown::TransposeDown(Note *note)
-   : Base("transpose_down")
+Action::Transform::TransposeUp::TransposeUp(Note *note)
+   : Base("transpose_up")
    , note(note)
 {}
 
 
 
 
-Action::Transform::TransposeDown::~TransposeDown()
+Action::Transform::TransposeUp::~TransposeUp()
 {}
 
 
 
 
-bool Action::Transform::TransposeDown::execute()
+bool Action::Transform::TransposeUp::execute()
 {
    if (!note) return false;
 
    std::vector<Note> single_note_as_array;
    single_note_as_array.push_back(*note);
 
-   ::Transform::Transpose transpose_transform(-1);
+   ::Transform::Transpose transpose_transform(1);
    single_note_as_array = transpose_transform.transform(single_note_as_array);
 
    *note = single_note_as_array.at(0);
